pyproject.toml
[tool.poetry]
name = "unicork-api"
version = "0.1.0"
description = ""
authors = ["donghhan <apensia914@gmail.com>"]
readme = "README.md"
packages = [{include = "unicork_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
